[IDLE-467] 채팅방 기능 연결 완료
