+++ feature/center/chatting/src/main/java/com/idle/center/chatting/CenterChattingFragment.kt
@@ -45,8 +45,11 @@ import com.idle.designsystem.compose.component.LoadingCircle
 import com.idle.designsystem.compose.foundation.CareTheme
 import com.idle.domain.model.auth.UserType
 import com.idle.domain.model.chat.ChatRoom
+import com.idle.domain.model.profile.CenterProfile
 import com.idle.domain.util.formatRelativeDateTime
 import com.idle.domain.util.formatUnReadNumber
+import com.idle.navigation.DeepLinkDestination
+import com.idle.navigation.NavigationEvent
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
@@ -57,6 +60,7 @@ internal class CenterChattingFragment : BaseComposeFragment() {
     override fun ComposeLayout() {
         fragmentViewModel.apply {
             val chatRoomList by chatRoomList.collectAsStateWithLifecycle()
+            val myProfile by myProfile.collectAsStateWithLifecycle()
 
             LifecycleEventEffect(Lifecycle.Event.ON_CREATE) {
                 getChatRoomList()
@@ -65,12 +69,9 @@ internal class CenterChattingFragment : BaseComposeFragment() {
 
             CenterChattingScreen(
                 chatRoomList = chatRoomList,
+                myProfile = myProfile,
                 navigateTo = {
-                    navigationHelper.navigateTo(
-                        com.idle.navigation.NavigationEvent.To(
-                            it
-                        )
-                    )
+                    navigationHelper.navigateTo(NavigationEvent.To(it))
                 },
             )
         }
@@ -80,7 +81,8 @@ internal class CenterChattingFragment : BaseComposeFragment() {
 @Composable
 internal fun CenterChattingScreen(
     chatRoomList: List<ChatRoom>?,
-    navigateTo: (com.idle.navigation.DeepLinkDestination) -> Unit,
+    myProfile: CenterProfile?,
+    navigateTo: (DeepLinkDestination) -> Unit,
 ) {
     val listState = rememberLazyListState()
 
@@ -137,6 +139,7 @@ internal fun CenterChattingScreen(
                         ) { chatRoom ->
                             ChatRoomItem(
                                 chatRoom = chatRoom,
+                                myProfile = myProfile,
                                 navigateTo = navigateTo,
                             )
                         }
@@ -158,7 +161,8 @@ internal fun CenterChattingScreen(
 @Composable
 internal fun ChatRoomItem(
     chatRoom: ChatRoom,
-    navigateTo: (com.idle.navigation.DeepLinkDestination) -> Unit,
+    myProfile: CenterProfile?,
+    navigateTo: (DeepLinkDestination) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     val backgroundColor = if (chatRoom.unReadMessageCount > 0) {
@@ -174,9 +178,9 @@ internal fun ChatRoomItem(
             .background(backgroundColor)
             .clickable {
                 navigateTo(
-                    com.idle.navigation.DeepLinkDestination.ChattingDetail(
+                    DeepLinkDestination.ChattingDetail(
                         chattingRoomId = chatRoom.id,
-                        receiverId = chatRoom.myId,
+                        receiverId = myProfile!!.centerId,
                         senderId = chatRoom.opponentId,
                         receiverUserType = UserType.WORKER.apiValue,
                     )
